+++ lib/src/applied_mathematics/probability_theory/numbers_generator.dart
@@ -1,5 +1,7 @@
 import 'dart:math';
 
+import 'package:meta/meta.dart';
+
 /// Generator of random numbers
 ///
 /// It isn't extend `Random` of `dart:math`, but provide functionality
@@ -19,23 +21,22 @@ class NumbersGenerator {
   double nextDouble({double from = 0, double to = 1}) =>
       r.nextDouble() * (to + 0.001 - from) + from;
 
-  /// Returns [count] of integer numbers
+  /// Returns iterable of integer numbers
   ///
   /// Generates `integer` number in range [from] - [to] inclusively.
   /// [from] should be less than [to].
-  Iterable<int> intIterableSync(int to, int count, {int from = 0}) sync* {
-    for (var i = 0; i < count; i++) {
+  Iterable<int> intIterableSync({@required int to, int from = 0}) sync* {
+    while (true) {
       yield nextInt(to, from: from);
     }
   }
 
-  /// Returns [count] of double numbers
+  /// Returns iterable of double numbers
   ///
   /// Generates `double` number in range [from] - [to] inclusively.
   /// [from] should be less than [to].
-  Iterable<double> doubleIterableSync(int count,
-      {double from = 0, double to = 1}) sync* {
-    for (var i = 0; i < count; i++) {
+  Iterable<double> doubleIterableSync({double from = 0, double to = 1}) sync* {
+    while (true) {
       yield nextDouble(from: from, to: to);
     }
   }