+++ test/extended_math_test.dart
@@ -1,7 +1,8 @@
 import 'package:extended_math/extended_math.dart';
 
 void main(List<String> args) {
-  final a = 3;
-  final b = 4;
-  print(hypot(3, 4)); // 5
+  final gen = Vector(<num>[4, 3]);
+  gen[1] = 8;
+  //gen.insert(1);
+  print(gen);
 }